+++ drivers/media/huawei/camera/platform/hi3650/sensor_common.c
@@ -1096,7 +1096,7 @@ fail:
 
 void hw_camdrv_msleep(unsigned int ms)
 {
-	ktime_t wait = ns_to_ktime(ms * 1000 * 1000);
+	ktime_t wait = ns_to_ktime((u64)ms * 1000 * 1000);
 	set_current_state(TASK_UNINTERRUPTIBLE);
 	schedule_hrtimeout(&wait, HRTIMER_MODE_REL);
 }
